fix segfault at ffi boundary when compiling in release mode
There is a reproducible segfault when dereferencing PathSet iterators
for the first time, but only when compiling in release mode.
According to my gdb experiments, the value of `set` here:
libstore/src/lib.rs#L154
is not the same as the value of `ps` here:
libstore/wrapper.hpp#L50
for some obscure reason.
Since it appears that a nix::PathSet is 48 bytes long, it is a bit
too big to be passed as an argument, and this triggers xmm instructions.
This is the only reasonnable^W explanation as to why this only triggers
in release mode.
So I decided to replace this with pointers both because it should be
more efficient as an argument and because I hoped that it would fix the
segfault, and it did apparently.
Well, dark magic.