make -O2 parallel
I tried more complicated stuff to avoid contention on the RwLock of the
graph, but it ended up being not much faster (3%) than this on nvme and
significantly slower on hdd.

Specifically what I tried: parallelizing readdir and stat and putting
(node index, inode, stat) into a mpsc channel and then modifying the
graph in one single thread. With various degrees of smartness to avoir
stating when possible.